+++ lib/protocol/url.rb
@@ -5,6 +5,7 @@
 
 require_relative "url/version"
 require_relative "url/error"
+require_relative "url/invalid_path_error"
 require_relative "url/pattern"
 require_relative "url/encoding"
 require_relative "url/form_data/nested"